[WIP] #21 - rewriteTests
Removed updateOrganization from tests as well as OrganizationOp, as I am still working on those.

Switched all field value testing to use new random string generator. Each time a test involving creating or updating is run, a new set of values are generated as needed and are used to test for success.