[WIP] #4 - webhookAPI
Added class WebhookUpdateInput. This had to be a seperate class from WebhookInput because createWebhook has two required parameters while updateWebhook contain only optional parameters.

In addition, changed tests for updateWebhook to accommodate the fact that add_events and remove_events cannot be used in the same method call. Comment was added to the code to document this fact.