[WIP] #11 - organizationAPI
Added tests for all parameters in Organization. The fields including dates in Joda DateTime objects are compared by converting to UTC time. The updated_at field is compared to the current time within a range of 5 seconds.

updateOrganization's return type was changed to Future[OrganizationDetail] from Future[Option[OrganizationDetail]]. This assumes the user will not enter an invalid organization or repo, and that if they do, it is their bug.